scheduler_ly_complete
